+++ linkerd/core/src/main/scala/io/buoyant/linkerd/Router.scala
@@ -96,7 +96,7 @@ trait RouterConfig {
   def servers: Seq[ServerConfig]
   def client: Option[ClientConfig]
 
-  var baseDtab: Option[Dtab] = None
+  var dtab: Option[Dtab] = None
   var failFast: Option[Boolean] = None
   var originator: Option[Boolean] = None
   var timeoutMs: Option[Int] = None
@@ -183,7 +183,7 @@ trait RouterConfig {
 
   @JsonIgnore
   def routerParams = (Stack.Params.empty + defaultBudget)
-    .maybeWith(baseDtab.map(dtab => RoutingFactory.BaseDtab(() => dtab)))
+    .maybeWith(dtab.map(dtab => RoutingFactory.BaseDtab(() => dtab)))
     .maybeWith(failFast.map(FailFastFactory.FailFast(_)))
     .maybeWith(originator.map(Originator.Param(_)))
     .maybeWith(timeoutMs.map(timeout => TimeoutFilter.Param(timeout.millis)))